macOS window manager integration and working full-screen
Full-screen on macOS was resulting in an unresponsive black screen. Addressed the issue by switching to the native macOS full-screen function (though had to disable titlebar hiding to allow for this, due to API limitations).

Also addressed various other scenarios through testing, such as launching in a maximized+full-screen state or minimizing a maximized window.

Still need to test on Gnome and Windows, and might do some cleanup at the end.